[TEST] Fix backwards compatibility tests for 2.x
Without a 2.0.0 GA release we can only test against 2.0.0-beta1 so this adds
a temporary hack for that.

`CompositeTestCluster#client` was broken because it tried to make the client
before there were external nodes to connect to. Now its created when the
cluster is fully started and recreated after a node upgrade.

`CompositeTestCluster#upgradeOneNode(Settings)` claimed that the upgraded
nodes would share the same data directory as the node that it was replacing
but didn't do anything for force that to happen, causing any test that didn't
use replicas to fail.

`FunctionScoreBackwardCompatibilityIT` relied on tf/idf information but didn't
set the number of shards to 1 or use dfs_query_then_fetch. Now it uses one
shard.

This also fixes the docs and reproduction printer to work with the way the
tests work now.

Not all of the backwards tests are passing, but 2/3 of them are.

Closes #13425