Wait longer when --debug-jvm is specified
Causes the build to wait for 10 minutes instead of 30 seconds
for Elasticsearch to startup before integration tests when
`--debug-jvm` is on the command line. 30 seconds is very short
when debugging locally because it can take a while for the
build to get to the point where it start starts Elasticsearch.