Grok runtime fields
Adds a new runtime field type called `grok` that executes a `grok`
flavored regular expression on the fly:
```
curl -XPOST -HContent-Type:application/json -uelastic:password 'localhost:9200/test/_search?pretty&error_trace' -d'{
  "runtime_mappings": {
    "http": {
      "type": "grok",
      "patterns": "%{IPORHOST:clientip} %{HTTPDUSER:ident} %{USER:auth} \\[%{TIMESTAMP_ISO8601:timestamp}\\] \"(?:%{WORD:verb} %{NOTSPACE:request}(?: HTTP/%{NUMBER:httpversion})?|%{DATA:rawrequest})\" %{NUMBER:response:int} (?:%{NUMBER:bytes:long}|-)",
      "field": "message"
    }
  },
  "query": { "range": { "http.bytes": { "gt": 5000 } } },
  "fields": ["http.*"]
}'
```

The central feature here is `patterns` which allows you to add one or
more regular expressions that are executed when running the search. The
pattern in the example above matches log lines that look like apache
logs:
```
252.0.0.0 - - [1998-04-30T14:31:27-05:00] "GET /images/hm_bg.jpg HTTP/1.0" 200 24736
```

This `grok` regex language is the same one we use for the grok ingest
processor and includes a fairly large built in "pattern bank" for things
like `IPORHOST` and `TIMESTAMP_ISO8601`. The syntax for extracting a
match is `%{PATTERN:name:type}` with the `:type` part being optional.

One `grok` runtime makes many sub-fields named after the extracted
portion of the pattern. So the example above makes `http.clientip` and
`http.bytes` and `http.verb` and so on. This is mostly to keep things
tidy and organized together. The optional `:type` portion configures the
kind of runtime field:

Grok Type | Runtime field type
----------|-------------------
   string | keyword
<no type> | keyword
  boolean | boolean
     long | long
      int | long
   double | double
    float | double

`grok` supports `float` and `int` but runtime fields don't so we
"promote" `float` to `double` and `int` to `long`. This is probably ok,
but you might see different floating point errors than you'd expect if
you declare a field like `%{NUMBER:voltage:float}` because we'll promote
the float to a double. Happy campers don't try to match floating point
numbers exactly.