Reproduce an error with def casting
We don't throw the right kind of exception when you combine the
auto-property stuff with def casting. The simplest way to get this in
production ES is to have a field declared of type `double` and then
have a script that does something like:
```
float f = dev['field'].value
```

`value`, being a `double`, doesn't fit into `float`. If you try this
with a `def` typed variable containing a `double` it throws a sensible
`ClassCastException`. But if you are trying to read an auto-property
you get this lovely thing:
```
Caused by: java.lang.invoke.WrongMethodTypeException: cannot convert MethodHandle(Doubles)double to (Object)float
```

As a general rule, `MethodHandle` is a deep, dark implementation detail
of painless and we should avoid having it bubble to the top of the
exception chain. Its fine if, like, its under a sensible
`ClassCastException` as the `cause`. But its difficult to reason about
as the sole cause of the exception unless you know a little about
`invokedynamic`. And we don't expect folks to know about it.