Weaken assertion on top_hits explain test
The tests for the explanation of top_hits asserted that the expanation
matches precise strings. Which it mostly does. But there seem to be edge
cases where a number in the string varies. We really don't care about
the number. This replaces the string assertion with a regex that doesn't
care about the number.