ESQL: Basic enrich-like lookup loading
This adds super basic way to perform a lookup-style LEFT JOIN thing.
It's *like* ENRICH, except it can use an index_mode=lookup index rather
than an ENRICH policy.

This intentionally forks some portion of the ENRICH infrastructure and
shares others. I *believe* it's the right part to fork and share, but
we'll see!